requirements.txt
Django==4.1.7
django-debug-toolbar==3.8.1
djangorestframework==3.14.0
djoser==2.1.0
python-dotenv==0.21.1